[Fix] 修正 施放Buff技能 玩家狀態效果詳細視窗未正確更新計時時間的錯誤

+++ Assets/script/CharacterStatusHint/CharacterStatusHintIntroSetting.cs
@@ -80,7 +80,7 @@ public class CharacterStatusHintIntroSetting : MonoBehaviour
     /// <param name="timer"></param>
     private void TimeSetting(object o, float timer)
     {
-        buffTimerText.text = string.Format("TM_TimeLeft".GetText(), timer.ToString()); ;
+        buffTimerText.text = string.Format("TM_TimeLeft".GetText(), ((int)timer).ToString()); ;
     }
 
     public void DestroyGameObject()
